refactor: change structure and add modules
