+++ src/interface/readline.js
@@ -1,0 +1,7 @@
+import readline from 'readline';
+
+
+export const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout
+  });
